# golang/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API untuk manajemen penjadwalan sekolah dengan Guided GA-GWO algorithm",
        "title": "Timetabling Backend API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:12000",
    "basePath": "/api",
    "paths": {
...
        "/relasi-guru-mapel": {
            "get": {
                "description": "Get a list of all relasi guru mapel",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RelasiGuruMapel"
                ],
                "summary": "Get all relasi guru mapel",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RelasiGuruMapel"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new relasi guru mapel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RelasiGuruMapel"
                ],
                "summary": "Create relasi guru mapel",
                "parameters": [
                    {
                        "description": "Relasi guru mapel data",
                        "name": "relasi",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RelasiGuruMapel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.RelasiGuruMapel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/relasi-guru-mapel/{id}": {
            "get": {
                "description": "Get a specific relasi guru mapel by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RelasiGuruMapel"
                ],
                "summary": "Get relasi guru mapel by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Relasi ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RelasiGuruMapel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing relasi guru mapel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RelasiGuruMapel"
                ],
                "summary": "Update relasi guru mapel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Relasi ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated relasi data",
                        "name": "relasi",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RelasiGuruMapel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RelasiGuruMapel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a relasi guru mapel by ID",
                "tags": [
                    "RelasiGuruMapel"
                ],
                "summary": "Delete relasi guru mapel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Relasi ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.RelasiGuruMapel": {
            "type": "object",
            "properties": {
                "durasi": {
                    "type": "integer"
                },
                "guru_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "mapel_id": {
                    "type": "integer"
                },
                "tingkatan": {
                    "type": "integer"
                }
            }
        },
...
    }
}